PM-8900 Apex Class and Trigger modifications
Added new Tokens in BaseApexLexer.g4 to match specific keywords
Added new Rules in the BaseApexParser.g4 to match the new keywords defined.
Overriding visitor rules so that we can parse the file and identify the specific values that need to be matched/searched.
Since only 1 rule can be process need to identify the correct order to place them in so custom-specific rules match before a default rule matches

Updates: PM-8900